Add support for Hystrix and Failsafe based Circuit Breaker TokenInfoRequestExecutors
This will address #53 and allow to have special TokenInfoRequestExecutor implementations which integrate with Hystrix and Failsafe.

This will not change any default behaviour of the library so far. It is expected for any configuration code (either auto configuration or manual configuration) to pick the correct implementation. Both Hystrix as well as Failsafe are marked as `optional` dependencies in order to not increase the dependency graph.